Create neighborhood functions

+++ src/bin.rs
@@ -4,8 +4,8 @@ use som::*;
 
 // Main function
 fn main() {
-    let item: Vec<i32> = vec![2,3,4];
-    let v: Vec<&Vec<i32>> = vec![&item, &item, &item];
+    let item: Vec<isize> = vec![2,3,4];
+    let v: Vec<&Vec<isize>> = vec![&item, &item, &item];
 
     let som = SOM::new((10, 10));
     som.train(&v, 100);
